+++ integration_tests/bchain/postdata_test.py
@@ -10,14 +10,15 @@ from trufflepig.testutils.pytest_fixtures import steem
 from trufflepig.bchain.poster import Poster
 
 
-@pytest.mark.skipif(config.PASSWORD is None, reason="needs posting key")
 def test_test_top10post(steem):
 
-    steem.wallet.unlock(config.PASSWORD)
+    if config.PASSWORD:
+        steem.wallet.unlock(config.PASSWORD)
 
     poster = Poster(steem=steem,
                     account=config.ACCOUNT,
-                    waiting_time=0.1)
+                    waiting_time=0.1,
+                    no_posting_key_mode=config.PASSWORD is None)
 
     posts = random_data.create_n_random_posts(10)
     df = pd.DataFrame(posts)
@@ -35,14 +36,15 @@ def test_test_top10post(steem):
     tbpd.comment_on_own_top_list(df, poster, permalink)
 
 
-@pytest.mark.skipif(config.PASSWORD is None, reason="needs posting key")
 def test_test_all_top_with_real_data(steem):
 
-    steem.wallet.unlock(config.PASSWORD)
+    if config.PASSWORD:
+        steem.wallet.unlock(config.PASSWORD)
 
     poster = Poster(steem=steem,
                     account=config.ACCOUNT,
-                    waiting_time=0.1)
+                    waiting_time=0.1,
+                    no_posting_key_mode=config.PASSWORD is None)
 
     df = tpbg.scrape_hour_data(steem, stop_after=10)
 
@@ -62,14 +64,15 @@ def test_test_all_top_with_real_data(steem):
                                   overview_permalink='jjj')
 
 
-@pytest.mark.skipif(config.PASSWORD is None, reason="needs posting key")
 def test_test_top20_vote_and_comment(steem):
 
-    steem.wallet.unlock(config.PASSWORD)
+    if config.PASSWORD:
+        steem.wallet.unlock(config.PASSWORD)
 
     poster = Poster(steem=steem,
                     account=config.ACCOUNT,
-                    waiting_time=0.1)
+                    waiting_time=0.1,
+                    no_posting_key_mode=config.PASSWORD is None)
 
     posts = random_data.create_n_random_posts(10)
     df = pd.DataFrame(posts)
@@ -78,8 +81,6 @@ def test_test_top20_vote_and_comment(steem):
 
     df = tppp.preprocess(df, ncores=1)
 
-    account = config.ACCOUNT
-
     tbpd.vote_and_comment_on_topK(df, poster, 'laida',
                                   overview_permalink='lll')
 
@@ -87,3 +88,29 @@ def test_test_top20_vote_and_comment(steem):
 @pytest.mark.skipif(config.PASSWORD is None, reason="needs posting key")
 def test_create_wallet(steem):
     tbpd.create_wallet(steem, config.PASSWORD, config.POSTING_KEY)
+
+
+def test_test_top_trending_post(steem):
+
+    if config.PASSWORD:
+        steem.wallet.unlock(config.PASSWORD)
+
+    poster = Poster(steem=steem,
+                    account=config.ACCOUNT,
+                    waiting_time=0.1,
+                    no_posting_key_mode=config.PASSWORD is None)
+
+    posts = random_data.create_n_random_posts(10)
+    df = pd.DataFrame(posts)
+    df['reward'] = df.reward
+    df['predicted_votes'] = df.votes
+
+    df = tppp.preprocess(df, ncores=1)
+
+    date = pd.datetime.utcnow().date()
+
+    tbpd.post_top_trending_list(df, poster, date,
+                                overview_permalink='iii',
+                                trufflepicks_permalink='kkk',
+                                steem_amount=10,
+                                sbd_amount=10)