+++ trufflepig/utils.py
@@ -293,3 +293,15 @@ def none_retry(f, retries=16, sleep_time=2):
             time.sleep(sleep_time)
         return None
     return wrapped
+
+
+def none_error_retry(f, retries=5, sleep_time=11, errors=(RPCError,),
+                not_log_errors=(PostDoesNotExist,)):
+    """Combines Error and None retry"""
+    return none_retry(error_retry(f,
+                                  retries=retries,
+                                  sleep_time=sleep_time,
+                                  errors=errors,
+                                  not_log_errors=not_log_errors),
+                      retries=retries,
+                      sleep_time=sleep_time)