+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * SQLite is now used to store blockchain data
 * New Poster class to allow for waiting
+* New posting of trending without bots
+* BUG FIX: Images and links are now properly fitlered
+* BUG FIX: Better displaying on busy!
 
 *0.6.1a* - 2018-03-22